requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
filelock==3.6.0
idna==3.3
numpy==1.21.5
onedrivedownloader==1.1.1
Pillow==9.0.1
PySocks==1.7.1
requests==2.27.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
